refactor project files and create split TabsScreen logic
