+++ .idea/gradle.xml
@@ -19,7 +19,6 @@
             <option value="$PROJECT_DIR$/features" />
             <option value="$PROJECT_DIR$/features/main-menu" />
             <option value="$PROJECT_DIR$/features/sign-up" />
-            <option value="$PROJECT_DIR$/navigation" />
           </set>
         </option>
       </GradleProjectSettings>
