+++ app/src/main/java/ru/fefu/pnexpert/App.kt
@@ -4,4 +4,4 @@ import android.app.Application
 import dagger.hilt.android.HiltAndroidApp
 
 @HiltAndroidApp
-class PnExpertApp : Application()
+class App : Application()
