--- app/src/main/java/ru/fefu/pnexpert/main/presentation/components/Event.kt
@@ -1,8 +1,0 @@
-package ru.fefu.pnexpert.main.presentation.components
-
-import androidx.compose.runtime.Composable
-
-@Composable
-fun Event() {
-
-}
