+++ app/src/main/java/ru/fefu/pnexpert/presentation/navigation/tabsscreen/entities/TabItems.kt
@@ -1,0 +1,12 @@
+package ru.fefu.pnexpert.presentation.navigation.tabsscreen.entities
+
+/**
+ * Place tab here in order which you want to they placed in app
+ */
+val bottomNavTabItems = listOf(
+    BottomNavTab.Services,
+    MarkedBottomNabTab.Notes,
+    BottomNavTab.Main,
+    BottomNavTab.History,
+    BottomNavTab.Profile
+)
