+++ app/src/main/java/ru/fefu/pnexpert/presentation/navigation/tabsscreen/navigation/NavBuilderExtention.kt
@@ -1,0 +1,8 @@
+package ru.fefu.pnexpert.presentation.navigation.tabsscreen.navigation
+
+import androidx.navigation.NavGraphBuilder
+import ru.fefu.pnexpert.presentation.navigation.tabsscreen.entities.BottomNavTab
+
+fun NavGraphBuilder.registerTabsScreen() {
+
+}
