+++ app/src/main/res/values/strings.xml
@@ -6,5 +6,4 @@
     <string name="services">Сервисы</string>
     <string name="note">Запись</string>
     <string name="events">Мероприятия</string>
-    <string name="back_button">Кнопка назад</string>
 </resources>
