+++ core/presentation/src/main/res/drawable/ic_arrow_back.xml
