--- navigation/src/main/java/ru/fefu/navigation/presentation/navigation/NavComponentRouter.kt
@@ -1,4 +1,0 @@
-package ru.fefu.navigation.presentation.navigation
-
-class NavComponentRouter {
-}
